study: CGLib proxy study
