+++ src/test/java/study/proxy/HelloTargetConcrete.java
@@ -1,0 +1,16 @@
+package study.proxy;
+
+public class HelloTargetConcrete {
+
+    public String sayHello(String name) {
+        return "Hello " + name;
+    }
+
+    public String sayHi(String name) {
+        return "Hi " + name;
+    }
+
+    public String sayThankYou(String name) {
+        return "Thank You " + name;
+    }
+}
